Add new class TextReader to support dataset union and columns rename
